+++ inappchat/src/main/java/com/livechatinc/inappchat/ChatWindowLifecycleObserver.java
@@ -1,5 +1,6 @@
 package com.livechatinc.inappchat;
 
+import android.content.ActivityNotFoundException;
 import android.net.Uri;
 import android.os.Build;
 
@@ -16,10 +17,14 @@ import androidx.lifecycle.MutableLiveData;
 import java.util.Collections;
 import java.util.List;
 
-public class ChatWindowLifecycleObserver implements DefaultLifecycleObserver {
+class ChatWindowLifecycleObserver implements DefaultLifecycleObserver {
 
-    public ChatWindowLifecycleObserver(@NonNull ActivityResultRegistry registry) {
+    public ChatWindowLifecycleObserver(
+            @NonNull ActivityResultRegistry registry,
+            @NonNull ActivityNotFoundCallback activityNotFoundCallback
+    ) {
         this.registry = registry;
+        this.activityNotFoundCallback = activityNotFoundCallback;
     }
 
     private final ActivityResultRegistry registry;
@@ -27,7 +32,7 @@ public class ChatWindowLifecycleObserver implements DefaultLifecycleObserver {
     @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
     private ActivityResultLauncher<String> getMultipleContent;
     private final MutableLiveData<List<Uri>> resultLiveData = new MutableLiveData<>();
-
+    private final ActivityNotFoundCallback activityNotFoundCallback;
 
     public void onCreate(@NonNull LifecycleOwner owner) {
         registerSingleContentContract(owner);
@@ -42,7 +47,9 @@ public class ChatWindowLifecycleObserver implements DefaultLifecycleObserver {
                 "chatWindowFileResultRegisterKey",
                 owner,
                 new ActivityResultContracts.GetContent(),
-                (file) -> resultLiveData.postValue(Collections.singletonList(file))
+                (file) -> resultLiveData.postValue(
+                        file != null ? Collections.singletonList(file) : Collections.emptyList()
+                )
         );
     }
 
@@ -58,15 +65,29 @@ public class ChatWindowLifecycleObserver implements DefaultLifecycleObserver {
 
     public void selectFile() {
         //TODO: consider using accept type from chrome client callback
-        getContent.launch("*/*");
+        try {
+            getContent.launch("*/*");
+        } catch (ActivityNotFoundException exception) {
+            activityNotFoundCallback.onActivityNotFoundException();
+            resultLiveData.postValue(Collections.emptyList());
+        }
     }
 
     @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
     public void selectFiles() {
-        getMultipleContent.launch("*/*");
+        try {
+            getMultipleContent.launch("*/*");
+        } catch (ActivityNotFoundException exception) {
+            activityNotFoundCallback.onActivityNotFoundException();
+            resultLiveData.postValue(Collections.emptyList());
+        }
     }
 
     public LiveData<List<Uri>> getResultLiveData() {
         return resultLiveData;
     }
+}
+
+interface ActivityNotFoundCallback {
+    void onActivityNotFoundException();
 }