+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Android ChangeLog
 
+## Version 2.4.0 - Sep 13th, 2024
+Migrate to AndroidX for managing Fragments
+Allow for sending multiple files at once on Android 21+
+
+**Breaking Change**
+Move away from deprecated `startActivityForResult` to `registerForActivityResult` to start file picker Activity and handle its result.
+
+Removed:
+* `onStartFilePickerActivity` callback from `ChatWindowEventsListener`
+* `onActivityResult` method from `ChatWindowView`
+Added:
+* `setUpAttachmentSupport` in `ChatWindowView`
+* `onFilePickerActivityNotFound` in`ChatWindowEventsListener`
+See File Sharing section in the docs for more details.
+
 ## Version 2.3.4 - Sep 1st, 2024
 Remove obsolete WebView popup from layout and prevent unnecessary view inflation